remove array type option from spin isospin function

+++ nuclear_qmc/wave_function/get_spin_isospin_indices/get_spin_isospin_indices.py
@@ -106,7 +106,7 @@ def get_isospin_exchange_index(particle_pairs, mass_number, proton_number, as_ja
         return package.array(exchange_indices)
 
 
-def get_system_arrays_pairs_triplets_spin_and_isospin(n_protons, n_neutrons, dtype=jnp.float64, as_jax_array=True,
+def get_system_arrays_pairs_triplets_spin_and_isospin(n_protons, n_neutrons, as_jax_array=True,
                                                       also_return_binary_representation=False):
     """Helper function returns necessary arrays associated with the nuclear system."""
     mass_number = n_protons + n_neutrons
